rrdtool plugin: use rrd_updatex_r with RRD_SKIP_PAST_UPDATES flag
This required to avoid data loss when timestamp not grow for any reasons,
(for example - duplicated or reordered UDP packet, or due to possible bug in
Collectd core) and buggy Collectd code was unable to check this properly.

This resolves cases which are logged with a message "illegal attempt to update
using time XXX when last update time is XXX (minimum one second step)".
Recent rrdtool (>=1.5.0) allows to skip only certain datapoints, not all data.

collectd/collectd#1058